feat(docker): add InvokeAI root variables
In this git diff, we see two new environment variables (`HOST_INVOKEAI_ROOT` and `APP_INVOKEAI_ROOT`) added to the `.docker/docker.app.env` file. This likely to set and manage the root directories for InvokeAI in the Docker container and the host machine.

Additionally, an unnecessary space was removed, but this is too minor to include in the commit message.

The commit message summarizes this change: feat(docker): add InvokeAI root variables